fixed SPI a wiring error

+++ push.sh
@@ -1,3 +1,8 @@
+if [[ $# -eq 0 ]] ; then
+    echo 'error: please enter a commit comment'
+    exit 1
+fi
+
 find . -type f -name "*.b#*" -exec rm -f {} \;
 find . -type f -name "*.s#*" -exec rm -f {} \;
 find . -type f -name "*.l#*" -exec rm -f {} \;
